Adding machers both for response and for content of the response.
Implemented with inspiration from rails controller specs.

The ResponseMatcher contains matchers for the response like should_be_redirect
XmlMatcher gives the tester abilitys to assert the content of the response

A extention to HttpResponse via the submit method, which ensures that the response
stays inside the same session. And gives the developer a Label() class and submits
the input field that the label points to. Se specs for more info.